requirement.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.5
decorator==5.1.1
Django==4.0.4
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
Faker==13.11.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
num-thai==0.0.5
numpy==1.23.1
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
roman==3.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
